incremental hash
