+++ src/Security/Cryptography/Hash/Blake/Blake2b.cs
@@ -7,9 +7,6 @@
 namespace CryptoHives.Foundation.Security.Cryptography.Hash;
 
 using System;
-using System.Buffers.Binary;
-using System.Numerics;
-using System.Runtime.CompilerServices;
 
 /// <summary>
 /// Computes the BLAKE2b hash for the input data.
@@ -28,7 +25,7 @@ using System.Runtime.CompilerServices;
 /// BLAKE2b supports an optional key for keyed hashing (MAC mode) with keys up to 64 bytes.
 /// </para>
 /// </remarks>
-public sealed partial class Blake2b : HashAlgorithm
+public sealed class Blake2b : HashAlgorithm
 {
     /// <summary>
     /// The maximum hash size in bits.
@@ -50,36 +47,9 @@ public sealed partial class Blake2b : HashAlgorithm
     /// </summary>
     public const int BlockSizeBytes = 128;
 
-    // Rounds of mixing
-    // private const int Rounds = 12;
-
-    // The size of the state buffer
-    private const int StateSize = 8;
-
-    // The size of the scratch buffers
-    private const int ScratchSize = 16;
-
-    // BLAKE2b IV constants (same as SHA-512)
-    private static readonly ulong[] IV =
-    [
-        0x6a09e667f3bcc908UL,
-        0xbb67ae8584caa73bUL,
-        0x3c6ef372fe94f82bUL,
-        0xa54ff53a5f1d36f1UL,
-        0x510e527fade682d1UL,
-        0x9b05688c2b3e6c1fUL,
-        0x1f83d9abfb41bd6bUL,
-        0x5be0cd19137e2179UL
-    ];
-
-    // Scalar state for non-AVX2 path and output extraction
-    private readonly SimdSupport _simdSupport;
-    private readonly ulong[] _state;
-    private readonly byte[] _buffer;
+    // Blake2b core state
+    private Blake2bState _core;
     private readonly byte[]? _key;
-    private readonly int _outputBytes;
-    private ulong _bytesCompressed;
-    private int _bufferLength;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="Blake2b"/> class with default output size (64 bytes).
@@ -125,16 +95,8 @@ public sealed partial class Blake2b : HashAlgorithm
                 $"Key size must be between 0 and {MaxKeySizeBytes} bytes.");
         }
 
-        _outputBytes = outputBytes;
         HashSizeValue = outputBytes * 8;
-        _state = new ulong[StateSize];
-        _buffer = new byte[BlockSizeBytes];
-
-        _simdSupport = SimdSupport.None;
-#if NET8_0_OR_GREATER
-        _simdSupport = simdSupport & SimdSupport;
-#endif
-
+        _core = new Blake2bState(simdSupport, outputBytes);
         if (key != null && key.Length > 0)
         {
             _key = new byte[key.Length];
@@ -200,163 +162,29 @@ public sealed partial class Blake2b : HashAlgorithm
     /// <returns>A new BLAKE2b instance configured for keyed hashing.</returns>
     internal static Blake2b CreateKeyed(SimdSupport simdSupport, int outputBytes, byte[] key) => new(simdSupport, outputBytes, key);
 
+#if NET8_0_OR_GREATER
+    /// <summary>
+    /// Gets the SIMD instruction sets supported by this algorithm on the current platform.
+    /// </summary>
+    internal static new SimdSupport SimdSupport => Blake2bState.SimdSupport;
+#endif
+
     /// <inheritdoc/>
     public override void Initialize()
     {
-        int keyLength = _key?.Length ?? 0;
-        ulong paramBlock = 0x01010000UL | ((ulong)keyLength << 8) | (uint)_outputBytes;
-
-        InitializeState(paramBlock);
-
-        _bytesCompressed = 0;
-        _bufferLength = 0;
-        ClearBuffer(_buffer);
-
-        // If keyed, the first block is the zero-padded key
-        if (_key != null && _key.Length > 0)
-        {
-            Array.Copy(_key, _buffer, _key.Length);
-            _bufferLength = BlockSizeBytes;
-        }
-    }
-
-    [MethodImpl(MethodImplOptionsEx.HotPath)]
-    private unsafe void Compress(byte* msgPtr, ulong* state, int bytesConsumed, bool isFinal)
-    {
-        _bytesCompressed += (ulong)bytesConsumed;
-
-#if NET8_0_OR_GREATER
-        if ((_simdSupport & SimdSupport.Avx2) != 0)
-        {
-            CompressAvx2(msgPtr, state, _bytesCompressed, isFinal);
-        }
-#if EXPERIMENTAL
-        else if ((_simdSupport & SimdSupport.Neon) != 0)
-        {
-            CompressNeon(msgPtr, state, _bytesCompressed, isFinal);
-        }
-#endif
-        else
-#endif
-        {
-            CompressScalar(msgPtr, state, _bytesCompressed, isFinal);
-        }
-    }
-
-    [MethodImpl(MethodImplOptionsEx.HotPath)]
-    private unsafe void Compress(ReadOnlySpan<byte> block, ulong* state, int bytesConsumed, bool isFinal)
-    {
-        fixed (byte* msgPtr = block)
-        {
-            Compress(msgPtr, state, bytesConsumed, isFinal);
-        }
-    }
-
-    [MethodImpl(MethodImplOptionsEx.HotPath)]
-    private unsafe void Compress(ReadOnlySpan<byte> block, int bytesConsumed, bool isFinal)
-    {
-        fixed (ulong* state = _state)
-        fixed (byte* msgPtr = block)
-        {
-            Compress(msgPtr, state, bytesConsumed, isFinal);
-        }
-    }
-
-    private void InitializeState(ulong paramBlock)
-    {
-        Array.Copy(IV, _state, StateSize);
-        _state[0] ^= paramBlock;
+        _core.Reset(_key);
     }
 
     /// <inheritdoc/>
-    protected override unsafe void HashCore(ReadOnlySpan<byte> source)
+    protected override void HashCore(ReadOnlySpan<byte> source)
     {
-        int offset = 0;
-
-        fixed (ulong* state = _state)
-        {
-            // If we have data in buffer, fill it first
-            if (_bufferLength > 0)
-            {
-                int toCopy = Math.Min(BlockSizeBytes - _bufferLength, source.Length);
-                source.Slice(0, toCopy).CopyTo(_buffer.AsSpan(_bufferLength));
-                _bufferLength += toCopy;
-                offset += toCopy;
-
-                // Only compress if buffer is full AND there's more data coming
-                // (we need to keep the last block for finalization)
-                if (_bufferLength == BlockSizeBytes && offset < source.Length)
-                {
-                    Compress(_buffer, state, BlockSizeBytes, false);
-                    _bufferLength = 0;
-                }
-            }
-
-            // Process full blocks, but always keep at least one block for finalization
-            fixed (byte* msgPtr = source)
-            {
-                while (offset + BlockSizeBytes < source.Length)
-                {
-                    Compress(msgPtr + offset, state, BlockSizeBytes, false);
-                    offset += BlockSizeBytes;
-                }
-            }
-        }
-
-        // Store remaining bytes in buffer
-        if (offset < source.Length)
-        {
-            int remaining = source.Length - offset;
-            if (_bufferLength == 0)
-            {
-                source.Slice(offset, remaining).CopyTo(_buffer.AsSpan());
-                _bufferLength = remaining;
-            }
-            else
-            {
-                // Buffer already has data, append to it
-                source.Slice(offset, remaining).CopyTo(_buffer.AsSpan(_bufferLength));
-                _bufferLength += remaining;
-            }
-        }
+        _core.Append(source);
     }
 
     /// <inheritdoc/>
     protected override bool TryHashFinal(Span<byte> destination, out int bytesWritten)
     {
-        if (destination.Length < _outputBytes)
-        {
-            bytesWritten = 0;
-            return false;
-        }
-
-        // Zero-pad the remaining buffer
-        _buffer.AsSpan(_bufferLength).Clear();
-
-        // Compress final block
-        Compress(_buffer, _bufferLength, true);
-
-        // Extract output
-        ExtractOutput(destination);
-
-        bytesWritten = _outputBytes;
-        return true;
-    }
-
-    private void ExtractOutput(Span<byte> destination)
-    {
-        int fullWords = _outputBytes / sizeof(UInt64);
-
-        BinarySpans.WriteUInt64LittleEndian(_state.AsSpan(0, fullWords), destination);
-
-        // Handle partial final word
-        int remainingBytes = _outputBytes % sizeof(UInt64);
-        if (remainingBytes > 0)
-        {
-            Span<byte> temp = stackalloc byte[sizeof(UInt64)];
-            BinaryPrimitives.WriteUInt64LittleEndian(temp, _state[fullWords]);
-            temp.Slice(0, remainingBytes).CopyTo(destination.Slice(fullWords * sizeof(UInt64)));
-        }
+        return _core.TryGetCurrentHash(destination, out bytesWritten);
     }
 
     /// <inheritdoc/>
@@ -364,131 +192,9 @@ public sealed partial class Blake2b : HashAlgorithm
     {
         if (disposing)
         {
-            Array.Clear(_state, 0, _state.Length);
-            ClearBuffer(_buffer);
-            if (_key != null)
-            {
-                ClearBuffer(_key);
-            }
+            _core.Dispose();
+            ClearBuffer(_key);
         }
         base.Dispose(disposing);
-    }
-
-    [SkipLocalsInit]
-    [MethodImpl(MethodImplOptionsEx.OptimizedLoop)]
-    private static unsafe void CompressScalar(byte* msgPtr, ulong* state, ulong bytesCompressed, bool isFinal)
-    {
-        // On little-endian platforms (all current .NET targets), the message bytes can be
-        // interpreted directly as ulong words via pointer cast — unaligned reads are handled
-        // correctly by all x86-64 and ARM64 hardware. On big-endian, we byte-swap into a
-        // scratch buffer via BinarySpans. BitConverter.IsLittleEndian is a JIT constant,
-        // so the dead branch and its stackalloc are eliminated entirely.
-        ulong* mr = (ulong*)msgPtr;
-        if (!BitConverter.IsLittleEndian)
-        {
-            ulong* scratch = stackalloc ulong[ScratchSize];
-            BinarySpans.ReadUInt64LittleEndian(msgPtr, scratch, ScratchSize);
-            mr = scratch;
-        }
-
-        // Load current hash state into 16 local working-vector variables.
-        // Pointer dereference — no bounds checks, no ref tracking overhead.
-        ulong v0 = state[0], v1 = state[1], v2 = state[2], v3 = state[3],
-              v4 = state[4], v5 = state[5], v6 = state[6], v7 = state[7];
-
-        // IV constants
-        ulong v8 = IV[0], v9 = IV[1], v10 = IV[2], v11 = IV[3],
-              v12 = IV[4] ^ bytesCompressed, v13 = IV[5],
-              v14 = isFinal ? ~IV[6] : IV[6], v15 = IV[7];
-
-        ulong m0 = mr[0], m1 = mr[1], m2 = mr[2], m3 = mr[3],
-              m4 = mr[4], m5 = mr[5], m6 = mr[6], m7 = mr[7],
-              m8 = mr[8], m9 = mr[9], m10 = mr[10], m11 = mr[11],
-              m12 = mr[12], m13 = mr[13], m14 = mr[14], m15 = mr[15];
-
-        bool step11and12 = false;
-        while (true)
-        {
-            // 10+2 rounds, fully unrolled — no loop counter, no Sigma table access at runtime.
-            // Round 0 — sigma: 0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15
-            G(ref v0, ref v4, ref v8, ref v12, m0, m1); G(ref v1, ref v5, ref v9, ref v13, m2, m3);
-            G(ref v2, ref v6, ref v10, ref v14, m4, m5); G(ref v3, ref v7, ref v11, ref v15, m6, m7);
-            G(ref v0, ref v5, ref v10, ref v15, m8, m9); G(ref v1, ref v6, ref v11, ref v12, m10, m11);
-            G(ref v2, ref v7, ref v8, ref v13, m12, m13); G(ref v3, ref v4, ref v9, ref v14, m14, m15);
-            // Round 1 — sigma: 14,10,4,8,9,15,13,6,1,12,0,2,11,7,5,3
-            G(ref v0, ref v4, ref v8, ref v12, m14, m10); G(ref v1, ref v5, ref v9, ref v13, m4, m8);
-            G(ref v2, ref v6, ref v10, ref v14, m9, m15); G(ref v3, ref v7, ref v11, ref v15, m13, m6);
-            G(ref v0, ref v5, ref v10, ref v15, m1, m12); G(ref v1, ref v6, ref v11, ref v12, m0, m2);
-            G(ref v2, ref v7, ref v8, ref v13, m11, m7); G(ref v3, ref v4, ref v9, ref v14, m5, m3);
-
-            if (step11and12) break;
-            step11and12 = true;
-
-            // Round 2 — sigma: 11,8,12,0,5,2,15,13,10,14,3,6,7,1,9,4
-            G(ref v0, ref v4, ref v8, ref v12, m11, m8); G(ref v1, ref v5, ref v9, ref v13, m12, m0);
-            G(ref v2, ref v6, ref v10, ref v14, m5, m2); G(ref v3, ref v7, ref v11, ref v15, m15, m13);
-            G(ref v0, ref v5, ref v10, ref v15, m10, m14); G(ref v1, ref v6, ref v11, ref v12, m3, m6);
-            G(ref v2, ref v7, ref v8, ref v13, m7, m1); G(ref v3, ref v4, ref v9, ref v14, m9, m4);
-            // Round 3 — sigma: 7,9,3,1,13,12,11,14,2,6,5,10,4,0,15,8
-            G(ref v0, ref v4, ref v8, ref v12, m7, m9); G(ref v1, ref v5, ref v9, ref v13, m3, m1);
-            G(ref v2, ref v6, ref v10, ref v14, m13, m12); G(ref v3, ref v7, ref v11, ref v15, m11, m14);
-            G(ref v0, ref v5, ref v10, ref v15, m2, m6); G(ref v1, ref v6, ref v11, ref v12, m5, m10);
-            G(ref v2, ref v7, ref v8, ref v13, m4, m0); G(ref v3, ref v4, ref v9, ref v14, m15, m8);
-            // Round 4 — sigma: 9,0,5,7,2,4,10,15,14,1,11,12,6,8,3,13
-            G(ref v0, ref v4, ref v8, ref v12, m9, m0); G(ref v1, ref v5, ref v9, ref v13, m5, m7);
-            G(ref v2, ref v6, ref v10, ref v14, m2, m4); G(ref v3, ref v7, ref v11, ref v15, m10, m15);
-            G(ref v0, ref v5, ref v10, ref v15, m14, m1); G(ref v1, ref v6, ref v11, ref v12, m11, m12);
-            G(ref v2, ref v7, ref v8, ref v13, m6, m8); G(ref v3, ref v4, ref v9, ref v14, m3, m13);
-            // Round 5 — sigma: 2,12,6,10,0,11,8,3,4,13,7,5,15,14,1,9
-            G(ref v0, ref v4, ref v8, ref v12, m2, m12); G(ref v1, ref v5, ref v9, ref v13, m6, m10);
-            G(ref v2, ref v6, ref v10, ref v14, m0, m11); G(ref v3, ref v7, ref v11, ref v15, m8, m3);
-            G(ref v0, ref v5, ref v10, ref v15, m4, m13); G(ref v1, ref v6, ref v11, ref v12, m7, m5);
-            G(ref v2, ref v7, ref v8, ref v13, m15, m14); G(ref v3, ref v4, ref v9, ref v14, m1, m9);
-            // Round 6 — sigma: 12,5,1,15,14,13,4,10,0,7,6,3,9,2,8,11
-            G(ref v0, ref v4, ref v8, ref v12, m12, m5); G(ref v1, ref v5, ref v9, ref v13, m1, m15);
-            G(ref v2, ref v6, ref v10, ref v14, m14, m13); G(ref v3, ref v7, ref v11, ref v15, m4, m10);
-            G(ref v0, ref v5, ref v10, ref v15, m0, m7); G(ref v1, ref v6, ref v11, ref v12, m6, m3);
-            G(ref v2, ref v7, ref v8, ref v13, m9, m2); G(ref v3, ref v4, ref v9, ref v14, m8, m11);
-            // Round 7 — sigma: 13,11,7,14,12,1,3,9,5,0,15,4,8,6,2,10
-            G(ref v0, ref v4, ref v8, ref v12, m13, m11); G(ref v1, ref v5, ref v9, ref v13, m7, m14);
-            G(ref v2, ref v6, ref v10, ref v14, m12, m1); G(ref v3, ref v7, ref v11, ref v15, m3, m9);
-            G(ref v0, ref v5, ref v10, ref v15, m5, m0); G(ref v1, ref v6, ref v11, ref v12, m15, m4);
-            G(ref v2, ref v7, ref v8, ref v13, m8, m6); G(ref v3, ref v4, ref v9, ref v14, m2, m10);
-            // Round 8 — sigma: 6,15,14,9,11,3,0,8,12,2,13,7,1,4,10,5
-            G(ref v0, ref v4, ref v8, ref v12, m6, m15); G(ref v1, ref v5, ref v9, ref v13, m14, m9);
-            G(ref v2, ref v6, ref v10, ref v14, m11, m3); G(ref v3, ref v7, ref v11, ref v15, m0, m8);
-            G(ref v0, ref v5, ref v10, ref v15, m12, m2); G(ref v1, ref v6, ref v11, ref v12, m13, m7);
-            G(ref v2, ref v7, ref v8, ref v13, m1, m4); G(ref v3, ref v4, ref v9, ref v14, m10, m5);
-            // Round 9 — sigma: 10,2,8,4,7,6,1,5,15,11,9,14,3,12,13,0
-            G(ref v0, ref v4, ref v8, ref v12, m10, m2); G(ref v1, ref v5, ref v9, ref v13, m8, m4);
-            G(ref v2, ref v6, ref v10, ref v14, m7, m6); G(ref v3, ref v7, ref v11, ref v15, m1, m5);
-            G(ref v0, ref v5, ref v10, ref v15, m15, m11); G(ref v1, ref v6, ref v11, ref v12, m9, m14);
-            G(ref v2, ref v7, ref v8, ref v13, m3, m12); G(ref v3, ref v4, ref v9, ref v14, m13, m0);
-        }
-
-        // Finalize: state[i] ^= v[i] ^ v[i+8]
-        state[0] ^= v0 ^ v8; state[1] ^= v1 ^ v9;
-        state[2] ^= v2 ^ v10; state[3] ^= v3 ^ v11;
-        state[4] ^= v4 ^ v12; state[5] ^= v5 ^ v13;
-        state[6] ^= v6 ^ v14; state[7] ^= v7 ^ v15;
-    }
-
-    /// <summary>
-    /// BLAKE2b mixing function G.
-    /// </summary>
-    [MethodImpl(MethodImplOptionsEx.HotPath)]
-    private static void G(ref ulong a, ref ulong b, ref ulong c, ref ulong d, ulong x, ulong y)
-    {
-        unchecked
-        {
-            a = a + b + x;
-            d = BitOperations.RotateRight(d ^ a, 32);
-            c = c + d;
-            b = BitOperations.RotateRight(b ^ c, 24);
-            a = a + b + y;
-            d = BitOperations.RotateRight(d ^ a, 16);
-            c = c + d;
-            b = BitOperations.RotateRight(b ^ c, 63);
-        }
     }
 }
